Replace PL_main_root with CvROOT(PL_main_cv), and similar for PL_main_start.
This makes PL_main_cv consistent with every other CV, in that CvROOT() and
CvSTART() are used to point to the root and start of its optree.
The discrepancy arises because PL_main_cv was only added in 5.001 (as part of
adding closures), whereas PL_main_root and PL_main_start date from 5.000.