Bring bleadperl up to parity with CPAN for version.pm
Please find attached a patch to bleadperl to bring up to date with the
CPAN release of version.pm.  This release was to deal primarily with an
edge case in numifying alpha decimal versions:

   https://rt.cpan.org/Ticket/Display.html?id=79259

I've also synchronized all of the other version.pm test cases, so that
the code in core is identical to the CPAN release.

Signed-off-by: David Golden <dagolden@cpan.org>